requirements.txt
pyTelegramBotAPI
pyyaml